soros_rag_chatbot-main/requirements.txt
pandas
numpy
scikit-learn
torch
transformers
sentence-transformers
langchain-community
chromadb
openpyxl
sentencepiece
yfinance
google-generativeai
streamlit